Customized test base for desktop only code

+++ src/main/java/com/projectname/qa/base/TestBase.java
@@ -71,12 +71,16 @@ public class TestBase {
 		e_driver.register(eventListener);
 		driver = e_driver;
 		
-		//driver.manage().window().maximize();
-		driver.manage().deleteAllCookies();
-		driver.manage().timeouts().pageLoadTimeout(TestUtil.PAGE_LOAD_TIMEOUT, TimeUnit.SECONDS);
-		driver.manage().timeouts().implicitlyWait(TestUtil.IMPLICIT_WAIT, TimeUnit.SECONDS);
+		if (Globalplatform.equalsIgnoreCase("desktop"))
+		{
+			//driver.manage().window().maximize();
+			driver.manage().deleteAllCookies();
+			driver.manage().timeouts().pageLoadTimeout(TestUtil.PAGE_LOAD_TIMEOUT, TimeUnit.SECONDS);
+			driver.manage().timeouts().implicitlyWait(TestUtil.IMPLICIT_WAIT, TimeUnit.SECONDS);
+			
+	        driver.get(TestBase.GlobalURL);
+		}
 
-        driver.get(GlobalURL);
 		return driver;
     }
